// front/src/types.ts
export type YearData = {
    fiscal_year: number;
    num_reg: number;
  };

export type FundsData = {
  fiscal_year: number;
  supplier_contract_amount_usd: number;
}

export type PracticeData = {
  procurement_category: string;
  num_projects: number;
}

export type PracticeFunds = {
  procurement_category: string;
  supplier_contract_amount_usd: number;
}